@@ -14,7 +14,7 @@ description: a simple version of suyu
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -47,6 +47,8 @@ dependencies:
   extended_image: ^1.3.0
   ionicons: ^0.0.3
   syncfusion_flutter_calendar: ^18.3.47
+  emoji_picker: ^0.1.0
+  flutter_slider_drawer: ^1.0.3
 
 dev_dependencies:
   flutter_test:
