@@ -2,12 +2,11 @@ library animated_splash;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:provider/provider.dart';
-import 'package:suyu_simple/provider/FontFamilyProvider.dart';
+import 'package:suyu_simple/common/ThemeColor.dart';
 
 Widget _home;
 Function _customFunction;
-String _imagePath;
+
 int _duration;
 Color _color;
 AnimatedSplashType _runfor;
@@ -18,8 +17,7 @@ Map<dynamic, Widget> _outputAndHome = {};
 
 class AnimatedSplash extends StatefulWidget {
   AnimatedSplash(
-      {@required String imagePath,
-      @required Widget home,
+      {@required Widget home,
       Function customFunction,
       int duration,
       AnimatedSplashType type,
@@ -27,12 +25,10 @@ class AnimatedSplash extends StatefulWidget {
       Map<dynamic, Widget> outputAndHome}) {
     assert(duration != null);
     assert(home != null);
-    assert(imagePath != null);
     _color = color;
     _home = home;
     _duration = duration;
     _customFunction = customFunction;
-    _imagePath = imagePath;
     _runfor = type;
     _outputAndHome = outputAndHome;
   }
@@ -43,24 +39,20 @@ class AnimatedSplash extends StatefulWidget {
 
 class _AnimatedSplashState extends State<AnimatedSplash>
     with SingleTickerProviderStateMixin {
-  AnimationController _animationController;
-  Animation _animation;
-
   @override
   void initState() {
     super.initState();
-    if (_duration < 1000) _duration = 2000;
-    _animationController = new AnimationController(
-        vsync: this, duration: Duration(milliseconds: 800));
-    _animation = Tween(begin: 0.5, end: 1.0).animate(CurvedAnimation(
-        parent: _animationController, curve: Curves.easeInCirc));
-    _animationController.forward();
+    if (_duration < 1000) _duration = 800;
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
   }
 
   @override
   void dispose() {
     super.dispose();
-    _animationController.reset();
   }
 
   navigator(home) {
@@ -88,65 +80,61 @@ class _AnimatedSplashState extends State<AnimatedSplash>
         backgroundColor: _color,
         body: Center(
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.end,
+            mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: <Widget>[
-              Hero(
-                  tag: "logoImg",
-                  child:
-                      SizedBox(height: 80.0, child: Image.asset(_imagePath))),
-              // ScaleTransition(
-              //   scale: _animation,
-              //   child: Hero(
-              //       tag: "logoImg",
-              //       child: SizedBox(
-              //           height: 150.0, child: Image.asset(_imagePath))
-              //           ),
-              // ),
-              SizedBox(
-                height: 50,
-              ),
-              Column(
-                children: [
-                  Text(
-                    "素",
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
-                      fontWeight: FontWeight.w300,
-                      fontSize: 36,
-                      fontFamily: 'kaitiFont',
+              Container(
+                margin: EdgeInsets.only(top: 10, bottom: 10),
+                padding:
+                    EdgeInsets.only(top: 15, bottom: 15, left: 5, right: 5),
+                decoration: BoxDecoration(
+                    // color: Colors.black,
                     ),
-                  ),
-                  Text(
-                    "语",
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
-                      fontWeight: FontWeight.w300,
-                      fontSize: 36,
-                      fontFamily: 'kaitiFont',
-                    ),
-                  ),
-                  Text(
-                    "评",
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Text(
+                      "素",
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                        color: Colors.black,
                         fontWeight: FontWeight.w300,
                         fontSize: 36,
-                        fontFamily: 'kaitiFont'),
-                  ),
-                  Text(
-                    "分",
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
-                      fontWeight: FontWeight.w300,
-                      fontSize: 36,
-                      fontFamily: 'kaitiFont',
+                        fontFamily: 'kaitiFont',
+                      ),
                     ),
-                  ),
-                  SizedBox(
-                    height: 50,
-                  ),
-                ],
+                    Text(
+                      "语",
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                        color: Colors.black,
+                        fontWeight: FontWeight.w300,
+                        fontSize: 36,
+                        fontFamily: 'kaitiFont',
+                      ),
+                    ),
+                    Text(
+                      "评",
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                          color: Colors.black,
+                          fontWeight: FontWeight.w300,
+                          fontSize: 36,
+                          fontFamily: 'kaitiFont'),
+                    ),
+                    Text(
+                      "分",
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                        color: Colors.black,
+                        fontWeight: FontWeight.w300,
+                        fontSize: 36,
+                        fontFamily: 'kaitiFont',
+                      ),
+                    ),
+                  ],
+                ),
               ),
             ],
           ),
