@@ -4,6 +4,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:suyu_simple/common/ThemeColor.dart';
 // import 'package:suyu_simple/data/ChatStream.dart';
 // import 'package:suyu_simple/model/ChatMessage.dart';
 import 'package:suyu_simple/provider/ChatProvider.dart';
@@ -22,21 +23,16 @@ class _RulePageState extends State<RulePage> {
   @override
   void initState() {
     super.initState();
-
-    print("初始化");
   }
 
   @override
   void deactivate() {
     super.deactivate();
-
-    print("离开");
   }
 
   @override
   void dispose() {
     super.dispose();
-
     inputController.dispose();
   }
 
@@ -44,40 +40,45 @@ class _RulePageState extends State<RulePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
-          Container(
-            height: 100,
-            child: AnimatedList(
-              initialItemCount:
-                  Provider.of<ChatProvider>(context, listen: false)
-                      .getTmpList()
-                      .length,
-              itemBuilder: (context, index, animation) {
-                return Text(Provider.of<ChatProvider>(context, listen: false)
-                    .getTmpList()
-                    .toString());
-              },
-            ),
+          Text(
+            "开发中",
+            style: TextStyle(fontSize: 100, color: Colors.grey[300]),
           ),
-          TextField(
-            controller: inputController,
-          ),
-          RaisedButton(
-              child: Text("发送"),
-              onPressed: () => DioUtils.request(
-                  "https://interface.meiriyiwen.com//article/random",
-                  method: DioUtils.GET)),
+          Container(),
+          //   height: 100,
+          //   child: AnimatedList(
+          //     initialItemCount:
+          //         Provider.of<ChatProvider>(context, listen: false)
+          //             .getTmpList()
+          //             .length,
+          //     itemBuilder: (context, index, animation) {
+          //       return Text(Provider.of<ChatProvider>(context, listen: false)
+          //           .getTmpList()
+          //           .toString());
+          //     },
+          //   ),
+          // ),
+          // TextField(
+          //   controller: inputController,
+          // ),
+          // RaisedButton(
+          //     child: Text("发送"),
+          //     onPressed: () => DioUtils.request(
+          //         "https://interface.meiriyiwen.com//article/random",
+          //         method: DioUtils.GET)),
         ],
       ),
       floatingActionButton: Column(
         mainAxisAlignment: MainAxisAlignment.end,
         children: <Widget>[
-          FloatingActionButton(
-            onPressed: () => print("1"),
-            tooltip: 'Increment',
-            child: Icon(Icons.add),
-          ),
+          // FloatingActionButton(
+          //   onPressed: () => print("1"),
+          //   tooltip: 'Increment',
+          //   child: Icon(Icons.add),
+          // ),
           SizedBox(height: 24.0),
         ],
       ),
